% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{boxcount}
\alias{boxcount}
\title{FUNCTION_TITLE}
\usage{
boxcount(x, verbose = FALSE)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{verbose}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
