% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcount.R
\name{boxcounting}
\alias{boxcounting}
\title{FUNCTION_TITLE}
\usage{
boxcounting(data, dist.breaks)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{dist.breaks}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
