% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrsum.R
\name{corrsum}
\alias{corrsum}
\title{FUNCTION_TITLE}
\usage{
corrsum(searcher, data, dist.breaks, min.actual.pairs = 2000,
  min.nr.samples.at.scale = 128, max.nr.samples.at.scale = 1024,
  batch.size = 32, verbose = FALSE)
}
\arguments{
\item{searcher}{PARAM_DESCRIPTION}

\item{data}{PARAM_DESCRIPTION}

\item{dist.breaks}{PARAM_DESCRIPTION}

\item{min.actual.pairs}{PARAM_DESCRIPTION, Default: 2000}

\item{min.nr.samples.at.scale}{PARAM_DESCRIPTION, Default: 128}

\item{max.nr.samples.at.scale}{PARAM_DESCRIPTION, Default: 1024}

\item{batch.size}{PARAM_DESCRIPTION, Default: 32}

\item{verbose}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
