% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{count_integers}
\alias{count_integers}
\title{Compute a histogram of positive integers.}
\usage{
count_integers(bins, max_bin = -1L)
}
\arguments{
\item{bins}{An integer vector with positive entries only.}

\item{max_bin}{An integer with the maximum expected bin in bins.}
}
\description{
Fast counting of positve integers in input vector bins.
}
