% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_searcher}
\alias{create_searcher}
\title{FUNCTION_TITLE}
\usage{
create_searcher(x, metric = "euclidian", exclude_samples = 0L,
  cluster_max_points = 64L, seed = 93453562L)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{metric}{PARAM_DESCRIPTION, Default: 'euclidian'}

\item{exclude_samples}{PARAM_DESCRIPTION, Default: 0}

\item{cluster_max_points}{PARAM_DESCRIPTION, Default: 64}

\item{seed}{PARAM_DESCRIPTION, Default: 93453562}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
Create ATRIA searcher (ball-tree).

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
