% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{data_set_radius}
\alias{data_set_radius}
\title{FUNCTION_TITLE}
\usage{
data_set_radius(searcher)
}
\arguments{
\item{searcher}{PARAM_DESCRIPTION}

\item{x}{An integer vector}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
Data set radius

This function returns a logical vector identifying if
there are leading NA, marking the leadings NA as TRUE and
everything else as FALSE.

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
