% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrsum.R
\name{distlimits}
\alias{distlimits}
\title{FUNCTION_TITLE}
\usage{
distlimits(searcher, data)
}
\arguments{
\item{searcher}{PARAM_DESCRIPTION}

\item{data}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stats]{quantile}}
}
