% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{henon}
\alias{henon}
\title{Leading NA}
\usage{
henon(length = 2000L, params = as.numeric(c()), transient = 5000L)
}
\arguments{
\item{x}{An integer vector}
}
\description{
This function returns a logical vector identifying if
there are leading NA, marking the leadings NA as TRUE and
everything else as FALSE.
}
