% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrsum.R
\name{logspace}
\alias{logspace}
\title{FUNCTION_TITLE}
\usage{
logspace(min.dist, max.dist, n.bins)
}
\arguments{
\item{min.dist}{PARAM_DESCRIPTION}

\item{max.dist}{PARAM_DESCRIPTION}

\item{n.bins}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
