% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{number_of_points}
\alias{number_of_points}
\title{FUNCTION_TITLE}
\usage{
number_of_points(searcher)
}
\arguments{
\item{searcher}{An exernal pointer to an ATRIA searcher.}

\item{searcher}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
Number of points,

Return number of points used to create the searcher.

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
