% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{release_searcher}
\alias{release_searcher}
\title{FUNCTION_TITLE}
\usage{
release_searcher(searcher)
}
\arguments{
\item{searcher}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
Release searcher

Destroy ATRIA searcher object and free all allocated memory.

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
