% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{search_k_neighbors}
\alias{search_k_neighbors}
\title{FUNCTION_TITLE}
\usage{
search_k_neighbors(searcher, k, query_points, exclude = matrix(),
  epsilon = 0)
}
\arguments{
\item{searcher}{PARAM_DESCRIPTION}

\item{k}{PARAM_DESCRIPTION}

\item{query_points}{PARAM_DESCRIPTION}

\item{exclude}{PARAM_DESCRIPTION, Default: matrix()}

\item{epsilon}{PARAM_DESCRIPTION, Default: 0}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
