% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{search_range}
\alias{search_range}
\title{FUNCTION_TITLE}
\usage{
search_range(searcher, radius, query_points, exclude = matrix())
}
\arguments{
\item{searcher}{PARAM_DESCRIPTION}

\item{radius}{PARAM_DESCRIPTION}

\item{query_points}{PARAM_DESCRIPTION}

\item{exclude}{PARAM_DESCRIPTION, Default: matrix()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
